requirements.txt
numpy==1.16.2
scipy==1.2.1
matplotlib==3.0.3
ipykernel==5.1.1
# openslide-python==1.1.2
opencv-python==4.1.0.25
scikit-learn==0.20.3
scikit-image==0.15.0
tensorflow-gpu==1.12
pillow==6.0.0
tqdm==4.35.0
h5py==2.10.0